+++ frontend/ui/src/app/app-routing.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import {HomeComponent} from "./components/pages/home/home.component";
 import {CalendarComponent} from "./components/pages/calendar/calendar.component";
+import {PersonComponent} from "./components/pages/society/person/person.component";
 
 const routes: Routes = [
   {
@@ -16,6 +17,10 @@ const routes: Routes = [
   {
     path: "calendar",
     component: CalendarComponent
+  },
+  {
+    path: "people",
+    component: PersonComponent
   }
 ];